repo fix and checklist addon

+++ src/main/java/fact/it/game_service/repository/GameRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface GameRepository extends JpaRepository<Game, Long> {
+public interface GameRepository extends JpaRepository<Game, Integer> {
     //Find all questions and order by theme.
     List<Game> findAllByOrderByTheme();
     //Find questions by theme
